+++ src/Resources/app/administration/src/module/ui-feed-export/page/ui-feed-export-index/index.js
@@ -1,0 +1,13 @@
+import template from './ui-feed-export-index.html.twig'
+
+const { Component } = Shopware;
+
+Component.register('ui-feed-export-index', {
+    template,
+
+    metaInfo() {
+        return {
+            title: this.$createTitle()
+        };
+    },
+});